# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.CheckSuspensionResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/entity.Suspension'
        type: array
      code:
        type: integer
      message:
        type: string
    type: object
  api.CreateBookRequest:
    properties:
      author:
        example: Robert W. Chambers
        maxLength: 100
        minLength: 5
        type: string
      description:
        example: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
          tempor incididunt ut labore et dolore magna aliqua.
        type: string
      tags:
        example:
        - horror
        - mystery
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
      title:
        example: The King in Yellow
        maxLength: 100
        minLength: 5
        type: string
      year:
        example: 1994
        minimum: 1
        type: integer
    required:
    - author
    - description
    - tags
    - title
    - year
    type: object
...
  api.CreateSuspensionRequest:
    properties:
      expires_in:
        description: Time in minutes
        minimum: 1
        type: integer
      reason:
        example: Bad behaviour
        type: string
      user_id:
        example: 21
        type: integer
    required:
    - expires_in
    - reason
    - user_id
    type: object
...
  api.DefaultResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.GetBookByIDResponse:
    properties:
      body:
        $ref: '#/definitions/entity.Book'
      code:
        type: integer
      message:
        type: string
    type: object
  api.GetBooksResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/entity.Book'
        type: array
      code:
        type: integer
      message:
        type: string
      meta:
        $ref: '#/definitions/util.Metadata'
    type: object
...
  api.UpdateBookRequest:
    properties:
      author:
        example: Robert W. Chambers
        maxLength: 100
        minLength: 5
        type: string
      description:
        example: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
          tempor incididunt ut labore et dolore magna aliqua.
        type: string
      tags:
        example:
        - horror
        - mystery
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
      title:
        example: The King in Yellow
        maxLength: 100
        minLength: 5
        type: string
      year:
        example: 1994
        minimum: 1
        type: integer
    type: object
...
  api.UpdateSuspensionRequest:
    properties:
      expires_in:
        description: Time in minutes
        minimum: 1
        type: integer
      reason:
        example: Bad behaviour
        type: string
    type: object
  entity.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      rating:
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
...
  entity.Suspension:
    properties:
      created_at:
        type: string
      expires_in:
        type: integer
      id:
        type: integer
      moderator_id:
        type: integer
      reason:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  util.Metadata:
    properties:
      current_page:
        type: integer
      first_page:
        type: integer
      last_page:
        type: integer
      page_size:
        type: integer
      totalRecords:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: dragon090986@gmail.com
    name: Kirill Shaforostov
  description: API for Library application
  title: Library API
  version: 0.0.1
paths:
  /books:
    get:
      parameters:
      - default: 1
        in: query
        name: page
        type: integer
      - default: 50
        description: Number of books inside of one page. Can range between 1-100
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - default: created_at
        description: The field that is used for sorting. Add prefix "-" to change
          direction
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetBooksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get books using pagination
      tags:
      - Books
  /books/{id}:
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Book info
          schema:
            $ref: '#/definitions/api.GetBookByIDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get book by id
      tags:
      - Books
...
  /books/delete/{id}:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete book by ID. Requires MODERATOR role or higher
      tags:
      - Books
  /books/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Book succesfully created
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new book. Requires MODERATOR role or higher
      tags:
      - Books
  /books/update/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Book succesfully updated
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update book by id. Requires MODERATOR role or higher
      tags:
      - Books
...
  /mod/suspensions/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.CreateSuspensionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User was successfully suspended
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Suspend user for some time. Requires MODERATOR role or higher
      tags:
      - Moderation
  /mod/suspensions/update/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Suspension ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UpdateSuspensionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: suspension was successfully updated
          schema:
            $ref: '#/definitions/api.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Modify suspension of user. Requires MODERATOR role or higher
      tags:
      - Moderation
...
  /users/suspensions/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/api.CheckSuspensionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Check if user has suspensions
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"